backend/requirements.lambda.txt
fastapi
mangum==0.17.0
pandas
pydantic
email-validator
python-dotenv
resend
